mark calculate and save to database

+++ PRN_ExamO_HE176160/PRN_ExamO_HE176160/Models/Result.cs
@@ -8,7 +8,7 @@ namespace PRN_ExamO_HE176160.Models
         public int ResultId { get; set; }
         public int? UserId { get; set; }
         public string? ExamId { get; set; }
-        public int? Marks { get; set; }
+        public double? Marks { get; set; }
 
         public virtual Exam? Exam { get; set; }
         public virtual User? User { get; set; }
